split up assert file
