+++ lighthouse-assert/assert/difference/interface.ts
@@ -1,0 +1,6 @@
+import { IDiff } from "../types";
+
+export interface IDifference {
+  getDiff: () => IDiff;
+  matchesExpectation: () => boolean;
+}
